Conditionally define NETWORK_CONTROLLER
The macro NETWORK_CONTROLLER should only be defined if it hasn't been defined before. Currently you'd have to change the library config.h code to use a different network controller. With this change the macro can be defined in the user space (.ino sketch possibly) and thus does not have to fiddle with library code.